handle mounts in recovery if the have only one argument
this allows the recovery to handle commands like 'mount system'
it is needed for devices with two different filesystem in
recovery.fstab like RFS and EXT4 cause the regular recovery
behaviour will only generate a fstab for mounts of fstype2
and ignore the other fstype. This will also enable things like

run_program("/sbin/busybox", "mount", "/system");

on those system.

+++ recovery.c
@@ -787,6 +787,11 @@ main(int argc, char **argv) {
             return nandroid_main(argc, argv);
         if (strstr(argv[0], "reboot"))
             return reboot_main(argc, argv);
+        if (strstr(argv[0], "mount") && argc == 2 && !strstr(argv[0], "umount"))
+        {
+			load_volume_table();
+            return ensure_path_mounted(argv[1]);
+		}
         if (strstr(argv[0], "poweroff")){
             return reboot_main(argc, argv);
         }
